Made binary representation function work on 64 bit ints and added some debugging/investigating code using that.
We found that there is no clear relationship between IOHIDEventField and CGEventField.

We should remove that debugging code displaying the binary representation of the different event field keys.

+++ Helper/Utility/Utility_Helper.h
@@ -12,7 +12,7 @@
 @interface Utility_Helper : NSObject
 + (CGEventRef)createEventWithValuesFromEvent:(CGEventRef)event;
 + (void)printEventFieldDifferencesBetween:(CGEventRef)event1 and:(CGEventRef)event2;
-+ (NSString *)binaryRepresentation:(int)value;
++ (NSString *)binaryRepresentation:(int64_t)value;
 + (CGPoint)getCurrentPointerLocation_flipped;
 @end
